By default, pass through UnknownFlag arguments with -Xclang
The reason being regular options are handled this way, -Xclang should do
the same, worst scenario is that preprocessor flags are included in hash
key and passed to the compiler. The former causes unnecessary cache
miss, the later may cause compiler warning. The chance of these events
are low since majorities of the flags are non-preprocessor flags and it
is not practical for sccache to know each of them.